token helper 리팩토링
