--- src/apps/domains/callback/helpers/token_helper.py
@@ -1,66 +1,0 @@
-from typing import Tuple
-
-import requests
-
-from apps.domains.callback.dtos import TokenData
-from apps.domains.callback.helpers.url_helper import UrlHelper
-from apps.domains.oauth2.models import Application
-from infra.configure.config import GeneralConfig
-
-
-class TokenHelper:
-    @staticmethod
-    def _generate_token_from_json(response_json: dict) -> Tuple[TokenData, TokenData]:
-        access_token = response_json.get('access_token', None)
-        access_token_expires_in = response_json.get('expires_in', None)
-
-        refresh_token = response_json.get('refresh_token', None)
-        refresh_token_expire_in = response_json.get('refresh_token_expires_in', None)
-
-        return TokenData(access_token, access_token_expires_in), TokenData(refresh_token, refresh_token_expire_in)
-
-    @classmethod
-    def _take_token(cls, client: Application, code: str, redirect_uri: str) -> Tuple[TokenData, TokenData]:
-        verify = not GeneralConfig.is_dev()
-        req = requests.post(
-            UrlHelper.get_oauth2_token_url(), data=cls._get_request_data(client, code, redirect_uri), verify=verify,
-        )
-        req.raise_for_status()
-
-        json = req.json()
-        return cls._generate_token_from_json(json)
-
-    @classmethod
-    def get_tokens(cls, client: Application, code: str, redirect_uri: str = None) -> Tuple[TokenData, TokenData]:
-        return cls._take_token(client, code, redirect_uri)
-
-    @staticmethod
-    def _get_request_data(client: Application, code: str, redirect_uri: str):
-        raise NotImplementedError()
-
-    @staticmethod
-    def _get_default_request_data(client: Application):
-        return {
-            'client_id': client.client_id,
-            'client_secret': client.client_secret,
-        }
-
-
-class TokenCodeHelper(TokenHelper):
-    @classmethod
-    def _get_request_data(cls, client: Application, code: str, redirect_uri: str) -> dict:
-        data = cls._get_default_request_data(client)
-        data['redirect_uri'] = redirect_uri
-        data['code'] = code
-        data['grant_type'] = 'authorization_code'
-        return data
-
-
-class TokenRefreshHelper(TokenHelper):
-    @classmethod
-    def _get_request_data(cls, client: Application, refresh_token: str, redirect_uri: str) -> dict:
-        data = cls._get_default_request_data(client)
-        data['refresh_token'] = refresh_token
-        data['grant_type'] = 'refresh_token'
-
-        return data
